Add files via upload
In this two cluster approaches are used: hierarchical clustering and K-means clustering. It is unsupervised learning technique for grouping related data points which shows same behaviour in the dataset regardless of the outcome. It organizes data into patterns that are simpler to understand and manipulate. There are many uses of cluster analysis such as market and image segmentation, search result grouping and social network analysis. The dataset contains customers id and feedback of user with scale from 0-10 and gives information whether respondent is professional or not. Customers age in years and there corresponding annual salary in pounds.